modified ConsoleUI

+++ ConsoleProject.Business/Interfaces/IEmployeeInterface.cs
@@ -5,7 +5,7 @@ namespace ConsoleProject.Business.Interfaces;
 public interface IEmployeeInterface
 {
     void CreateEmployee(Employee employee);
-    void UpdateEmployee(int employeeId, string name, string surname, decimal salary, int departmentId);
+    void UpdateEmployee(int employeeId, string name, string surname, double salary, int departmentId);
     void DeleteEmployee(int id);
     Employee GetByIdEmployee(int id);
     List<Employee> GetAllEmployee();
